fix: Resolve compilation errors and enhance Gantt service functionality
- Fix Entity Builder pattern issues by implementing @SuperBuilder
- Resolve DTO field name inconsistencies (predecessorTaskId -> predecessorId)
- Add missing GanttService methods (addTaskDependency, removeTaskDependency, etc.)
- Implement GanttChartDto.Statistics inner class for project analytics
- Update test files to match corrected DTO field names and method signatures
- Add React components foundation for future frontend migration
- Include sample data migration script for development

+++ src/test/java/com/liam/gantt/service/impl/ProjectServiceImplTest.java
@@ -365,7 +365,7 @@ class ProjectServiceImplTest {
         ProjectResponseDto responseWithProgress = ProjectResponseDto.builder()
                 .id(1L)
                 .name("테스트 프로젝트")
-                .averageProgress(BigDecimal.valueOf(75)) // (100 + 50) / 2
+                .averageProgress(75.0) // (100 + 50) / 2
                 .build();
 
         given(projectRepository.findById(1L)).willReturn(Optional.of(testProject));
